商品削除機能の実装_ご指摘箇所の修正v1

+++ app/controllers/items_controller.rb
@@ -37,6 +37,12 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
+
+    unless current_user && current_user.id == item.user_id
+      redirect_to root_path, alert: "編集権限がありません"
+      return
+    end
+
     item.destroy
     redirect_to root_path
   end
